move from_dict to Task
