+++ classes/ArchivedTask.py
@@ -31,3 +31,21 @@ class ArchivedTask(Task):
         }
 
         return res
+
+def from_dict(data):
+    if "archived_date" in data:
+        task = ArchivedTask()
+        task.archived_date = datetime.date.fromisoformat(data["archived_date"])
+    else:
+        task = Task(increment=False)
+    task.task_id = data["task_id"]
+    task.task_name = data["task_name"]
+    task.owner = data["owner"]
+    task.contributors = data["contributors"]
+    task.creation_date = datetime.date.fromisoformat(data["creation_date"])
+    task.due_date = datetime.date.fromisoformat(data["due_date"])
+    task.status = data["status"]
+    task.description = data["description"]
+    task.comments = data["comments"]
+
+    return task
